Int64s.

+++ zego/search.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Search_Results struct {
-    Count            int `json:"count"`
+    Count            int64 `json:"count"`
     NextPage         string `json:"next_page"`
     PrevPage         string `json:"prev_page"`
     Results          []*Result `json:"results"`
@@ -16,7 +16,7 @@ type Result struct {
     Name             string `json:"name"`
     CreatedAt        string `json:"created_at"`
     UpdatedAt        string `json:"updated_at"`
-    Id               int `json:"id"`
+    Id               int64 `json:"id"`
     ResultType       string `json:"result_type"`
     Url              string `json:"url"`
 }
